Query images that have titles that match a certain pattern
Switched to using an instantiation of the class. The
theme_mod_header_image action is run six times in the Twenty Fourteen
theme. It doesn't make sense to run a potentially expensive, impossible
to cache query six times so using an instantiation to allow caching the
image after the first query.